Renewed version of Air Purifier & Integrated version for Mask Box
Renewed version of Air Purifier by Minji Bae

Integrated Mask Box code with face detection and OpenAPI data processing by Luwis Choi
