+++ Mask_Box_Pi/README.md
@@ -1,0 +1,5 @@
+pip3 install pandas
+
+pip3 install xlrd
+
+sudo apt install fonts-unfonts-core
